refactor: separate Knowledge Base into independent node
- BREAKING: Split functionality into two specialized nodes
  - Docutray: Document processing (Convert/Identify operations)
  - Docutray Knowledge Base: RAG-based knowledge base search
- Create DocutrayKnowledgeBase node with dedicated search operation
- Update package.json to register both nodes
- Enhance documentation structure for multi-node architecture
- Follow n8n best practices for separation of concerns
