+++ CLAUDE.md
@@ -13,20 +13,22 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 
 ## Architecture
 
-This is an n8n community node package for Docutray integration using the n8n nodes API version 1. The package provides OCR document processing and document type identification capabilities through the Docutray API. The project follows n8n's standard node development patterns:
+This is an n8n community node package for Docutray integration using the n8n nodes API version 1. The package provides two specialized nodes: one for OCR document processing and document type identification, and another for knowledge base search using RAG technology. The project follows n8n's standard node development patterns:
 
 ### Project Structure
 - `/nodes/` - Contains node implementations (*.node.ts files)
-  - Expected: Docutray node with Convert, Identify, and Knowledge Base Search operations
+  - `Docutray/` - Document processing node (Convert and Identify operations)
+  - `DocutrayKnowledgeBase/` - Knowledge base search node (Search operation)
 - `/credentials/` - Contains credential type definitions (*.credentials.ts files)
-  - Expected: DocutrayApi credentials for API key authentication
+  - Expected: DocutrayApi credentials for API key authentication (shared between both nodes)
 - `/dist/` - Compiled output (generated by build process)
 - `gulpfile.js` - Handles icon copying during build
 
 ### Node Implementation Pattern
 Nodes implement the `INodeType` interface with:
 - `description: INodeTypeDescription` - Defines UI, properties, credentials, and API defaults
-- Resource/Operation pattern for organizing functionality (Document and Knowledge Base resources)
+- **Docutray node**: Resource/Operation pattern (Document resource with Convert/Identify operations)
+- **DocutrayKnowledgeBase node**: Direct operation pattern (Search operation only)
 - TypeScript compilation to CommonJS modules targeting ES2019
 
 ### Docutray API Operations
